pptools/table_final.py: Do not assume |pf|**3 slowdown for unsolved functions
Now (also in the context of nfevs) it seems that this introduces an
absolutely wild error. Instead, also in context of a suggestion by N.
Hansen, we report median slowdown across solved functions and portion of
solved functions relative to the baseline.

This still makes algorithms very difficult to compare.  However, the
rate of portfolio-solved functions should soar when we start to reuse
single-algorithm mdat files.